Unify decompress preprcessor
