+++ src/preprocessor/decompress.rs
@@ -1,0 +1,312 @@
+// Copyright 2020-2021, The Tremor Team
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use super::Preprocessor;
+use crate::errors::Result;
+use simd_json::ValueAccess;
+use std::io::{self, Read};
+use tremor_script::Value;
+
+#[derive(Clone, Default, Debug)]
+struct Gzip {}
+impl Preprocessor for Gzip {
+    fn name(&self) -> &str {
+        "gzip"
+    }
+
+    fn process(&mut self, _ingest_ns: &mut u64, data: &[u8]) -> Result<Vec<Vec<u8>>> {
+        use libflate::gzip::MultiDecoder;
+        let mut decoder = MultiDecoder::new(data)?;
+        let mut decompressed = Vec::new();
+        decoder.read_to_end(&mut decompressed)?;
+        Ok(vec![decompressed])
+    }
+}
+
+#[derive(Clone, Default, Debug)]
+struct Zlib {}
+impl Preprocessor for Zlib {
+    fn name(&self) -> &str {
+        "zlib"
+    }
+
+    fn process(&mut self, _ingest_ns: &mut u64, data: &[u8]) -> Result<Vec<Vec<u8>>> {
+        use libflate::zlib::Decoder;
+        let mut decoder = Decoder::new(data)?;
+        let mut decompressed = Vec::new();
+        decoder.read_to_end(&mut decompressed)?;
+        Ok(vec![decompressed])
+    }
+}
+
+#[derive(Clone, Default, Debug)]
+struct Xz2 {}
+impl Preprocessor for Xz2 {
+    fn name(&self) -> &str {
+        "xz2"
+    }
+
+    fn process(&mut self, _ingest_ns: &mut u64, data: &[u8]) -> Result<Vec<Vec<u8>>> {
+        use xz2::read::XzDecoder as Decoder;
+        let mut decoder = Decoder::new(data);
+        let mut decompressed = Vec::new();
+        decoder.read_to_end(&mut decompressed)?;
+        Ok(vec![decompressed])
+    }
+}
+
+#[derive(Clone, Default, Debug)]
+struct Snappy {}
+impl Preprocessor for Snappy {
+    fn name(&self) -> &str {
+        "snappy"
+    }
+
+    fn process(&mut self, _ingest_ns: &mut u64, data: &[u8]) -> Result<Vec<Vec<u8>>> {
+        use snap::read::FrameDecoder;
+        let mut rdr = FrameDecoder::new(data);
+        let decompressed_len = snap::raw::decompress_len(data)?;
+        let mut decompressed = Vec::with_capacity(decompressed_len);
+        io::copy(&mut rdr, &mut decompressed)?;
+        Ok(vec![decompressed])
+    }
+}
+
+#[derive(Clone, Default, Debug)]
+struct Lz4 {}
+impl Preprocessor for Lz4 {
+    fn name(&self) -> &str {
+        "lz4"
+    }
+
+    fn process(&mut self, _ingest_ns: &mut u64, data: &[u8]) -> Result<Vec<Vec<u8>>> {
+        use lz4::Decoder;
+        let mut decoder = Decoder::new(data)?;
+        let mut decompressed = Vec::new();
+        decoder.read_to_end(&mut decompressed)?;
+        Ok(vec![decompressed])
+    }
+}
+
+#[derive(Clone, Debug, Default)]
+struct Zstd {}
+impl Preprocessor for Zstd {
+    fn name(&self) -> &str {
+        "ztd"
+    }
+    fn process(&mut self, _ingest_ns: &mut u64, data: &[u8]) -> Result<Vec<Vec<u8>>> {
+        let decoded: Vec<u8> = zstd::decode_all(data)?;
+        Ok(vec![decoded])
+    }
+}
+
+#[derive(Clone, Default, Debug)]
+struct Fingerprinted {}
+impl Preprocessor for Fingerprinted {
+    fn name(&self) -> &str {
+        "fingerprinted"
+    }
+
+    fn process(&mut self, _ingest_ns: &mut u64, data: &[u8]) -> Result<Vec<Vec<u8>>> {
+        let r = match data.get(0..6) {
+            Some(&[0x1f, 0x8b, _, _, _, _]) => {
+                use libflate::gzip::Decoder;
+                let mut decoder = Decoder::new(data)?;
+                let mut decompressed = Vec::new();
+                decoder.read_to_end(&mut decompressed)?;
+                decompressed
+            }
+            // ZLib magic headers
+            Some(&[0x78, 0x01 | 0x5e | 0x9c | 0xda, _, _, _, _]) => {
+                use libflate::zlib::Decoder;
+                let mut decoder = Decoder::new(data)?;
+                let mut decompressed = Vec::new();
+                decoder.read_to_end(&mut decompressed)?;
+                decompressed
+            }
+            Some(&[0xfd, b'7', b'z', b'X', b'Z', 0x00]) => {
+                use xz2::read::XzDecoder as Decoder;
+                let mut decoder = Decoder::new(data);
+                let mut decompressed = Vec::new();
+                decoder.read_to_end(&mut decompressed)?;
+                decompressed
+            }
+            // Some(b"sNaPpY") => {
+            Some(&[0xff, _, _, _, _, _]) => {
+                use snap::read::FrameDecoder;
+                let mut rdr = FrameDecoder::new(data);
+                let decompressed_len = snap::raw::decompress_len(data)?;
+                let mut decompressed = Vec::with_capacity(decompressed_len);
+                io::copy(&mut rdr, &mut decompressed)?;
+                decompressed
+            }
+            Some(&[0x04, 0x22, 0x4D, 0x18, _, _]) => {
+                use lz4::Decoder;
+                let mut decoder = Decoder::new(data)?;
+                let mut decompressed = Vec::new();
+                decoder.read_to_end(&mut decompressed)?;
+                decompressed
+            }
+            // Zstd Magic : 0xFD2FB528 (but little endian)
+            Some(&[0x28, 0xb5, 0x2f, 0xfd, _, _]) => zstd::decode_all(data)?,
+            _ => data.to_vec(),
+        };
+        Ok(vec![r])
+    }
+}
+
+pub(crate) struct Decompress {
+    codec: Box<dyn Preprocessor>,
+}
+impl Decompress {
+    pub(crate) fn from_config(config: Option<&Value>) -> Result<Self> {
+        let codec: Box<dyn Preprocessor> = match config.get_str("algorithm") {
+            Some("gzip") => Box::new(Gzip::default()),
+            Some("zlib") => Box::new(Zlib::default()),
+            Some("xz2") => Box::new(Xz2::default()),
+            Some("snappy") => Box::new(Snappy::default()),
+            Some("lz4") => Box::new(Lz4::default()),
+            Some("zstd") => Box::new(Zstd::default()),
+            Some("autodetect") | None => Box::new(Fingerprinted::default()),
+            Some(other) => return Err(format!("Unknown decompression algorithm: {other}").into()),
+        };
+        Ok(Decompress { codec })
+    }
+}
+impl Preprocessor for Decompress {
+    fn name(&self) -> &str {
+        "fingerprinted"
+    }
+    fn process(&mut self, ingest_ns: &mut u64, data: &[u8]) -> Result<Vec<Vec<u8>>> {
+        self.codec.process(ingest_ns, data)
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use crate::postprocessor::Postprocessor;
+    use crate::preprocessor::Preprocessor;
+
+    macro_rules! assert_simple_symmetric {
+        ($internal:expr, $which:ident, $magic:expr) => {
+            fn decode_magic(data: &[u8]) -> &'static str {
+                match data.get(0..6) {
+                    Some(&[0x1f, 0x8b, _, _, _, _]) => "gzip",
+                    Some(&[0x78, _, _, _, _, _]) => "zlib",
+                    Some(&[0xfd, b'7', b'z', _, _, _]) => "xz2",
+                    Some(b"sNaPpY") => "snap",
+                    Some(&[0xff, 0x6, 0x0, 0x0, _, _]) => "snap",
+                    Some(&[0x04, 0x22, 0x4d, 0x18, _, _]) => "lz4",
+                    Some(&[0x28, 0xb5, 0x2f, 0xfd, _, _]) => "zstd",
+                    _ => "fail/unknown",
+                }
+            } // Assert pre and post processors have a sensible default() ctor
+            let mut pre = super::$which::default();
+            let mut post = crate::postprocessor::$which::default();
+
+            // Fake ingest_ns and egress_ns
+            let mut ingest_ns = 0_u64;
+            let egress_ns = 1_u64;
+
+            let r = post.process(ingest_ns, egress_ns, $internal);
+            let ext = &r?[0];
+            let ext = ext.as_slice();
+            // Assert actual encoded form is as expected ( magic code only )
+            assert_eq!($magic, decode_magic(&ext));
+
+            let r = pre.process(&mut ingest_ns, &ext);
+            let out = &r?[0];
+            let out = out.as_slice();
+            // Assert actual decoded form is as expected
+            assert_eq!(&$internal, &out);
+            // assert empty finish, no leftovers
+            assert!(pre.finish(None)?.is_empty())
+        };
+    }
+
+    macro_rules! assert_decompress {
+        ($internal:expr, $which:ident, $magic:expr) => {
+            // Assert pre and post processors have a sensible default() ctor
+            let mut pre = Decompress::from_config(None)?;
+            let mut outbound = crate::postprocessor::$which::default();
+
+            // Fake ingest_ns and egress_ns
+            let mut ingest_ns = 0_u64;
+            let egress_ns = 1_u64;
+
+            let r = outbound.process(ingest_ns, egress_ns, $internal);
+            let ext = &r?[0];
+            let ext = ext.as_slice();
+            // Assert actual encoded form is as expected ( magic code only )
+            assert_eq!($magic, decode_magic(&ext));
+
+            let r = pre.process(&mut ingest_ns, &ext);
+            let out = &r?[0];
+            let out = out.as_slice();
+            // Assert actual decoded form is as expected
+            assert_eq!(&$internal, &out);
+
+            // empty finish, no leftovers
+            assert!(pre.finish(None)?.is_empty());
+        };
+    }
+
+    #[test]
+    fn test_gzip() -> Result<()> {
+        let int = "snot".as_bytes();
+        assert_simple_symmetric!(int, Gzip, "gzip");
+        assert_decompress!(int, Gzip, "gzip");
+        Ok(())
+    }
+
+    #[test]
+    fn test_zlib() -> Result<()> {
+        let int = "snot".as_bytes();
+        assert_simple_symmetric!(int, Zlib, "zlib");
+        assert_decompress!(int, Zlib, "zlib");
+        Ok(())
+    }
+
+    #[test]
+    fn test_snappy() -> Result<()> {
+        let int = "snot".as_bytes();
+        assert_simple_symmetric!(int, Snappy, "snap");
+        assert_decompress!(int, Snappy, "snap");
+        Ok(())
+    }
+
+    #[test]
+    fn test_xz2() -> Result<()> {
+        let int = "snot".as_bytes();
+        assert_simple_symmetric!(int, Xz2, "xz2");
+        assert_decompress!(int, Xz2, "xz2");
+        Ok(())
+    }
+
+    #[test]
+    fn test_lz4() -> Result<()> {
+        let int = "snot".as_bytes();
+        assert_simple_symmetric!(int, Lz4, "lz4");
+        assert_decompress!(int, Lz4, "lz4");
+        Ok(())
+    }
+    #[test]
+    fn test_zstd() -> Result<()> {
+        let int = "snot".as_bytes();
+        assert_simple_symmetric!(int, Zstd, "zstd");
+        assert_decompress!(int, Zstd, "zstd");
+        Ok(())
+    }
+}
